# packages/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "xiaohei-erp",
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register/send-code": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Send Register Code",
        "operationId": "send_register_code_api_auth_register_send_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhoneVerificationCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneVerificationCodeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Logout",
        "operationId": "logout_api_auth_logout_post",
        "parameters": [
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Me",
        "operationId": "me_api_auth_me_get",
        "parameters": [
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/extension/auth": {
      "post": {
        "tags": [
          "extension"
        ],
        "summary": "Extension Auth",
        "operationId": "extension_auth_api_extension_auth_post",
        "parameters": [
          {
            "name": "erp_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Erp Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtensionAuthRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtensionAuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthUser": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "subscription_status": {
            "type": "string",
            "title": "Subscription Status"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "username",
          "role",
          "status",
          "subscription_status"
        ],
        "title": "AuthUser"
      },
...
      "ExtensionAuthRequest": {
        "properties": {
          "store_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Id"
          }
        },
        "type": "object",
        "title": "ExtensionAuthRequest"
      },
      "ExtensionAuthResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "store_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Id"
          }
        },
        "type": "object",
        "required": [
          "token",
          "expires_at",
          "store_id"
        ],
        "title": "ExtensionAuthResponse"
      },
...
      "FeatureFlagResponse": {
        "properties": {
          "feature_key": {
            "type": "string",
            "title": "Feature Key"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "feature_key",
          "enabled",
          "source"
        ],
        "title": "FeatureFlagResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "session": {
            "$ref": "#/components/schemas/SessionInfoResponse"
          }
        },
        "type": "object",
        "required": [
          "session"
        ],
        "title": "LoginResponse"
      },
      "LogoutResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          }
        },
        "type": "object",
        "title": "LogoutResponse"
      },
...
      "PhoneVerificationCodeRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 8,
            "title": "Phone"
          }
        },
        "type": "object",
        "required": [
          "phone"
        ],
        "title": "PhoneVerificationCodeRequest"
      },
      "PhoneVerificationCodeResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At"
          },
          "debug_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debug Code"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "expires_at"
        ],
        "title": "PhoneVerificationCodeResponse"
      },
...
      "RegisterRequest": {
        "properties": {
          "company_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 2,
            "title": "Company Name"
          },
          "phone": {
            "type": "string",
            "maxLength": 32,
            "minLength": 8,
            "title": "Phone"
          },
          "verification_code": {
            "type": "string",
            "maxLength": 12,
            "minLength": 4,
            "title": "Verification Code"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "company_name",
          "phone",
          "verification_code",
          "password"
        ],
        "title": "RegisterRequest"
      },
...
      "SessionInfoResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles"
          },
          "feature_flags": {
            "items": {
              "$ref": "#/components/schemas/FeatureFlagResponse"
            },
            "type": "array",
            "title": "Feature Flags"
          },
          "subscription_status": {
            "type": "string",
            "title": "Subscription Status"
          }
        },
        "type": "object",
        "required": [
          "user",
          "roles",
          "feature_flags",
          "subscription_status"
        ],
        "title": "SessionInfoResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}